Refactor Dictionary page with lazy-loaded components and improved structure
